requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
eval_type_backport==0.2.2
Flask==3.1.1
flower==2.0.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
kombu==5.5.4
markdownify==0.13.1
MarkupSafe==3.0.2
mistralai==1.8.1
more-itertools==10.7.0
munch==4.0.0
numpy==1.26.3
openai==1.90.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
playwright==1.54.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jobspy==1.1.80
pytz==2025.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlacodegen==3.1.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tls-client==1.0.1
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
